tokenizacao com passport jwt e logica de services/controllers atrelado ao user logado adicionadas

+++ kanban-field/src/boards/boards.service.ts
@@ -12,21 +12,21 @@ export class BoardsService {
   constructor(@InjectModel(Board.name) private boardModel: Model<BoardDocument>,
               private columnsService: ColumnsService) {}
 
-  async create(createBoardDto: CreateBoardDto) {
+  async create(createBoardDto: CreateBoardDto, userId: string) {
     try {
-      const board = new this.boardModel(createBoardDto);
+      const board = new this.boardModel({...createBoardDto, responsible: userId});
       return await board.save();
     } catch (error) {
       throw new Error(`Falha ao criar o quadro: ${error.message}`);
     }
   }
 
-  async findAll() {
+  async findAll(userId: string) {
     try {
-      const boards = await this.boardModel.find();
+      const boards = await this.boardModel.find({ responsible: userId });
   
       const boardsWithColumms = await Promise.all(boards.map(async (board) => {
-        board.columns = await this.columnsService.find({ board: board._id });
+        board.columns = await this.columnsService.find({ board: board._id }, userId);
         return board;
       }));
   
@@ -36,21 +36,21 @@ export class BoardsService {
     }
   }
 
-  async findOne(id: string) {
-    const board = await this.boardModel.findById(id);
+  async findOne(id: string, userId: string) {
+    const board = await this.boardModel.findById({_id: id, responsible: userId});
     
     if (!board) {
       throw new NotFoundException('Quadro não encontrado');
     }
 
-    board.columns = await this.columnsService.find({ board: id });
+    board.columns = await this.columnsService.find({ board: id }, userId );
     
     return board; // retorna o quadro e as colunas pertencentes a ele
   }
 
-  async update(id: string, updateBoardDto: UpdateBoardDto) {
+  async update(id: string, updateBoardDto: UpdateBoardDto, userId: string) {
     const board = await this.boardModel.findByIdAndUpdate(
-      id, updateBoardDto, { new: true }
+      {_id: id, responsible: userId }, updateBoardDto, { new: true }
     )
 
     if (!board) {
@@ -60,8 +60,8 @@ export class BoardsService {
     return board
   }
 
-  async remove(id: string) {
-    const board = await this.boardModel.findByIdAndDelete(id);
+  async remove(id: string, userId: string) {
+    const board = await this.boardModel.findByIdAndDelete({_id: id, responsible: userId});
     
     if (!board) {
       throw new NotFoundException('Quadro não encontrado');
